% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility-sim.R
\name{ate_calc}
\alias{ate_calc}
\title{Calculate Average Treatment Effect}
\usage{
ate_calc(t, y)
}
\arguments{
\item{t}{treatment assignment (must be 0 or 1)}

\item{y}{outcome (must be 0 or 1)}
}
\description{
returns the average treatment effect given treatment assignment in a
randomised scenario
}
