% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility-sim.R
\name{simulate_fragility}
\alias{simulate_fragility}
\title{Simulate Fragility Index}
\usage{
simulate_fragility(
  intervention_mort,
  power = 0.8,
  control_mort,
  n = NULL,
  sims = 1000,
  alpha = 0.05,
  permit_negative = FALSE
)
}
\arguments{
\item{intervention_mort}{treatment arm mortality: scalar \link{0, 1}}

\item{power}{desired statistical power: scalar \link{0, 1}}

\item{control_mort}{control arm mortality: scalar \link{0, 1}}

\item{n}{manual selection of subjects in total (overrides other settings)}

\item{sims}{number of simulations to run: default 1000}

\item{alpha}{pre-fixed alpha level to determine a "positive" finding.}

\item{permit_negative}{allow negative fragility index: default FALSE}
}
\description{
This is the workhorse function of the study. It calculates the appropriate
number of subjects for a given effect size (specified as a control and
intervention mortality), power and alpha. Simulations are repeated a default
1000 times (set by sims). Alternately, one can override this behaviour by
specifying the number of subjects directly (set by N).
}
\details{
The \code{permit_negative} refers to whether or not you want the FI to return
values for when the starting p value is already > 0.05.

A tibble is returned with list columns containing the original treatment
allocations and outcomes. A number of other metrics are returned, some
observed, some defined apriori. A column specifying \code{reverse_effect}
identifies cases where the reverse effect was present (i.e. harm observed in
the treatment arm). These cases should either be excluded, or flipped for
them to be valid and interpretable.
}
